updated package statements after move to Raptor test package
