+++ contribs/minibus/src/test/java/org/matsim/contrib/minibus/raptor/SameOriginDestinationRaptorTest.java
@@ -17,7 +17,7 @@
  *                                                                         *
  * *********************************************************************** */
 
-package org.matsim.contrib.minibus;
+package org.matsim.contrib.minibus.raptor;
 
 import org.junit.Rule;
 import org.junit.Test;
